IGNITE-20811 IGNITE-17700 Refactoring DataInput hierarchy
