+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/wal/serializer/RecordV1Serializer.java
@@ -31,14 +31,13 @@ import org.apache.ignite.internal.pagemem.wal.record.WALRecord.RecordType;
 import org.apache.ignite.internal.processors.cache.persistence.tree.io.CacheVersionIO;
 import org.apache.ignite.internal.processors.cache.persistence.wal.ByteBufferBackedDataInput;
 import org.apache.ignite.internal.processors.cache.persistence.wal.ByteBufferExpander;
+import org.apache.ignite.internal.processors.cache.persistence.wal.Crc32CheckingDataInput;
 import org.apache.ignite.internal.processors.cache.persistence.wal.SegmentEofException;
 import org.apache.ignite.internal.processors.cache.persistence.wal.WALPointer;
 import org.apache.ignite.internal.processors.cache.persistence.wal.WalSegmentTailReachedException;
 import org.apache.ignite.internal.processors.cache.persistence.wal.crc.FastCrc;
-import org.apache.ignite.internal.processors.cache.persistence.wal.io.FileInput;
 import org.apache.ignite.internal.processors.cache.persistence.wal.io.SegmentFileInputFactory;
 import org.apache.ignite.internal.processors.cache.persistence.wal.io.SegmentIO;
-import org.apache.ignite.internal.processors.cache.persistence.wal.io.SimpleFileInput;
 import org.apache.ignite.internal.processors.cache.persistence.wal.record.HeaderRecord;
 import org.apache.ignite.internal.processors.cache.persistence.wal.serializer.io.RecordIO;
 import org.apache.ignite.internal.processors.cache.version.GridCacheVersion;
@@ -221,12 +220,17 @@ public class RecordV1Serializer implements RecordSerializer {
     }
 
     /** {@inheritDoc} */
+    @Override public RecordDataSerializer recordDataSerializer() {
+        return dataSerializer;
+    }
+
+    /** {@inheritDoc} */
     @Override public void writeRecord(WALRecord rec, ByteBuffer buf) throws IgniteCheckedException {
         writeWithCrc(rec, buf, recordIO);
     }
 
     /** {@inheritDoc} */
-    @Override public WALRecord readRecord(FileInput in0, WALPointer expPtr) throws IOException, IgniteCheckedException {
+    @Override public WALRecord readRecord(ByteBufferBackedDataInput in0, WALPointer expPtr) throws IOException, IgniteCheckedException {
         return readWithCrc(in0, expPtr, recordIO);
     }
 
@@ -354,7 +358,7 @@ public class RecordV1Serializer implements RecordSerializer {
     /**
      * Reads record from file {@code in0} and validates CRC of record.
      *
-     * @param in0 File input.
+     * @param in0 Byte buffer input.
      * @param expPtr Expected WAL pointer for record. Used to validate actual position against expected from the file.
      * @param reader Record reader I/O interface.
      * @return WAL record.
@@ -362,13 +366,13 @@ public class RecordV1Serializer implements RecordSerializer {
      * @throws IgniteCheckedException If it's unable to read record.
      */
     static WALRecord readWithCrc(
-        FileInput in0,
+        ByteBufferBackedDataInput in0,
         WALPointer expPtr,
         RecordIO reader
     ) throws EOFException, IgniteCheckedException {
         long startPos = -1;
 
-        try (SimpleFileInput.Crc32CheckingFileInput in = in0.startRead(skipCrc)) {
+        try (Crc32CheckingDataInput in = new Crc32CheckingDataInput(in0, skipCrc)) {
             startPos = in0.position();
 
             WALRecord res = reader.readWithHeaders(in, expPtr);
@@ -386,7 +390,7 @@ public class RecordV1Serializer implements RecordSerializer {
             long size = -1;
 
             try {
-                size = in0.io().size();
+                size = in0.size();
             }
             catch (IOException ignore) {
                 // It just for information. Fail calculate file size.
